+++ src/client/App/index.tsx
@@ -13,13 +13,13 @@ import {
   usePersistUsername,
 } from '../usernameLocalStorage';
 import { useShareDB } from './useShareDB';
-import { VZLeft } from './VZLeft';
-import { VZMiddle } from './VZMiddle';
-import { VZRight } from './VZRight';
+import { VZLeft } from '../VZLeft';
+import { VZMiddle } from '../VZMiddle';
+import { VZRight } from '../VZRight';
 import {
   VZCodeContext,
   VZCodeProvider,
-} from './VZCodeContext';
+} from '../VZCodeContext';
 import './style.scss';
 
 // Instantiate the Prettier worker.
