+++ src/client/VZMiddle.tsx
@@ -1,10 +1,10 @@
 import { useContext } from 'react';
-import { SplitPaneResizeContext } from '../SplitPaneResizeContext';
-import { TabList } from '../TabList';
-import { CodeEditor } from '../CodeEditor';
-import { CodeErrorOverlay } from '../CodeErrorOverlay';
-import { PresenceNotifications } from '../PresenceNotifications';
-import { AIAssistWidget } from '../AIAssistWidget';
+import { SplitPaneResizeContext } from './SplitPaneResizeContext';
+import { TabList } from './TabList';
+import { CodeEditor } from './CodeEditor';
+import { CodeErrorOverlay } from './CodeErrorOverlay';
+import { PresenceNotifications } from './PresenceNotifications';
+import { AIAssistWidget } from './AIAssistWidget';
 import { VZCodeContext } from './VZCodeContext';
 
 // The middle portion of the VZCode environment, containing:
