+++ src/index.ts
@@ -7,38 +7,18 @@ export type {
   ShareDBDoc,
   VZCodeContent,
 } from './types';
-export type { EditorCache } from './client/useEditorCache';
 export type {
   ThemeLabel,
   ThemeOption,
-} from './client/themes';
-
-export { VZCodeProvider } from './client/App/VZCodeContext';
-export { VZSidebar } from './client/VZSidebar';
-export { VZSettings } from './client/VZSettings';
-export { VZResizer } from './client/VZResizer';
-export { VZLeft } from './client/App/VZLeft';
-export { VZMiddle } from './client/App/VZMiddle';
+  EditorCache,
+} from './client';
 
 export {
+  VZCodeProvider,
+  VZResizer,
+  VZLeft,
+  VZMiddle,
   SplitPaneResizeProvider,
   SplitPaneResizeContext,
-} from './client/SplitPaneResizeContext';
-export { TabList } from './client/TabList';
-export { CodeEditor } from './client/CodeEditor';
-export { usePrettier } from './client/usePrettier';
-export { CodeErrorOverlay } from './client/CodeErrorOverlay';
-export { PresenceNotifications } from './client/PresenceNotifications';
-export { useFileCRUD } from './client/useFileCRUD';
-export { useSubmitOperation } from './client/useSubmitOperation';
-export {
-  vzReducer,
-  createInitialState,
-} from './client/vzReducer';
-export { useActions } from './client/useActions';
-export { useOpenDirectories } from './client/useOpenDirectories';
-export {
-  defaultTheme,
-  useDynamicTheme,
-} from './client/themes';
-export { useEditorCache } from './client/useEditorCache';
+  useSubmitOperation,
+} from './client';
